Remove all remains of Boneless-II; Boneless-III is now the only one.
