+++ boneless/arch/instr.py
@@ -1,155 +1,190 @@
-import types
+import abc
+import parse
 
-from .opcode import *
-
-
-__all__ = [
-    "R0", "R1", "R2", "R3", "R4", "R5", "R6", "R7",
-    "ADD", "ADDI", "AND", "CMP", "ILL", "J", "JAL", "JC", "JE", "JNC", "JNE", "JNO", "JNS", "JNZ",
-    "JO", "JR", "JS", "JSGE", "JSGT", "JSLE", "JSLT", "JUGE", "JUGT", "JULE", "JULT", "JZ",
-    "LD", "LDI", "LDX", "MOV", "MOVA", "MOVH", "MOVI", "MOVL", "NOP", "OR", "ROT", "ROL", "ROR",
-    "SLL", "SRA", "SRL", "ST", "STI", "STX", "SUB", "SUBI", "XCHG", "XOR",
-    "L", "assemble",
-]
+from . import mc
 
 
-def A_FORMAT(opcode, optype, rd, ra, rb):
-    assert rd in range(8) and ra in range(8) and rb in range(8)
-    return (((opcode & 0b11111) << 11) |
-            ((    rd &   0b111) <<  8) |
-            ((    ra &   0b111) <<  5) |
-            ((    rb &   0b111) <<  2) |
-            ((optype &    0b11) <<  0))
-
-def S_FORMAT(opcode, optype, rd, ra, amt):
-    assert rd in range(8) and ra in range(8) and amt in range(16)
-    return (((opcode & 0b11111) << 11) |
-            ((    rd &   0b111) <<  8) |
-            ((    ra &   0b111) <<  5) |
-            ((   amt &  0b1111) <<  1) |
-            ((optype &     0b1) <<  0))
-
-def M_FORMAT(opcode, rsd, ra, off):
-    assert rsd in range(8) and ra in range(8)
-    if isinstance(off, str):
-        return lambda resolve: M_FORMAT(opcode, rsd, ra, resolve(off))
-    assert -16 <= off <= 15
-    return (((opcode & 0b11111) << 11) |
-            ((   rsd &   0b111) <<  8) |
-            ((    ra &   0b111) <<  5) |
-            ((   off & 0b11111) <<  0))
-
-def I_FORMAT(opcode, rsd, imm, u=False):
-    assert rsd in range(8)
-    if isinstance(imm, str):
-        return lambda resolve: I_FORMAT(opcode, rsd, resolve(imm), u)
-    assert ((not u and -128 <= imm <= 127) or
-            (u and imm in range(256)))
-    return (((opcode & 0b11111) << 11) |
-            ((   rsd &   0b111) <<  8) |
-            ((   imm &    0xff) <<  0))
-
-def C_FORMAT(opcode, off):
-    if isinstance(off, str):
-        return lambda resolve: C_FORMAT(opcode, resolve(off))
-    assert -1024 <= off <= 1023
-    return (((opcode & 0b11111) << 11) |
-            ((   off &   0x7ff) <<  0))
+__all__ = ["Reg", "Imm3AL", "Imm3SR", "Imm5", "Imm8", "Imm13", "Instr"]
 
 
-R0, R1, R2, R3, R4, R5, R6, R7 = range(8)
+def _signed(value, bits):
+    if value & (1 << (bits - 1)):
+        return value | (-1 << bits)
+    return value
 
 
-def NOP ():            return [A_FORMAT(OPCODE_LOGIC,   OPTYPE_AND,  0,  0,  0)]
+class Reg(mc.Operand):
+    format    = "R{:d}"
 
-def AND (rd, ra, rb):  return [A_FORMAT(OPCODE_LOGIC,   OPTYPE_AND, rd, ra, rb)]
-def OR  (rd, ra, rb):  return [A_FORMAT(OPCODE_LOGIC,   OPTYPE_OR,  rd, ra, rb)]
-def XOR (rd, ra, rb):  return [A_FORMAT(OPCODE_LOGIC,   OPTYPE_XOR, rd, ra, rb)]
+    @classmethod
+    def prepare(cls, value):
+        value = int(value)
+        if value not in range(0, 8):
+            raise ValueError("Register operand must be one of R0 to R7")
+        return value
 
-def ADD (rd, ra, rb):  return [A_FORMAT(OPCODE_ARITH,   OPTYPE_ADD, rd, ra, rb)]
-def SUB (rd, ra, rb):  return [A_FORMAT(OPCODE_ARITH,   OPTYPE_SUB, rd, ra, rb)]
-def CMP (    ra, rb):  return [A_FORMAT(OPCODE_ARITH,   OPTYPE_CMP,  0, ra, rb)]
+    def __repr__(self):
+        return str(self)
 
-def SLL (rd, ra, amt): return [S_FORMAT(OPCODE_SHIFT_L, OPTYPE_SLL, rd, ra, amt)]
-def ROT (rd, ra, amt): return [S_FORMAT(OPCODE_SHIFT_L, OPTYPE_ROT, rd, ra, amt % 16)]
-def ROL (rd, ra, amt): return [S_FORMAT(OPCODE_SHIFT_L, OPTYPE_ROT, rd, ra, amt % 16)]
-def ROR (rd, ra, amt): return [S_FORMAT(OPCODE_SHIFT_L, OPTYPE_ROT, rd, ra, (16 - amt) % 16)]
-
-def SRL (rd, ra, amt): return [S_FORMAT(OPCODE_SHIFT_R, OPTYPE_SRL, rd, ra, amt)]
-def SRA (rd, ra, amt): return [S_FORMAT(OPCODE_SHIFT_R, OPTYPE_SRA, rd, ra, amt)]
-def MOV (rd, rs):      return [S_FORMAT(OPCODE_SHIFT_L, OPTYPE_SLL, rd, rs,   0)]
-
-def LD  (rd, ra, off): return [M_FORMAT(OPCODE_LD,   rd, ra, off)]
-def ST  (rs, ra, off): return [M_FORMAT(OPCODE_ST,   rs, ra, off)]
-def LDX (rd, ra, off): return [M_FORMAT(OPCODE_LDX,  rd, ra, off)]
-def STX (rs, ra, off): return [M_FORMAT(OPCODE_STX,  rs, ra, off)]
-
-def ADDI(rd, imm):     return [I_FORMAT(OPCODE_ADDI, rd, +imm)]
-def SUBI(rd, imm):     return [I_FORMAT(OPCODE_ADDI, rd, -imm)]
-
-def MOVL(rd, imm):     return [I_FORMAT(OPCODE_MOVL, rd, imm, u=True)]
-def MOVH(rd, imm):     return [I_FORMAT(OPCODE_MOVH, rd, imm, u=True)]
-
-def MOVA(rd, off):     return [I_FORMAT(OPCODE_MOVA, rd, off)]
-def LDI (rd, off):     return [I_FORMAT(OPCODE_LDI,  rd, off)]
-def STI (rs, off):     return [I_FORMAT(OPCODE_STI,  rs, off)]
-def JAL (rd, off):     return [I_FORMAT(OPCODE_JAL,  rd, off)]
-def JR  (rd, off):     return [I_FORMAT(OPCODE_JR,   rd, off)]
-
-def J   (off):         return [C_FORMAT(OPCODE_J,    off)]
-
-def JNZ (off):         return [C_FORMAT(OPCODE_JNZ,  off)]
-def JZ  (off):         return [C_FORMAT(OPCODE_JZ,   off)]
-def JNS (off):         return [C_FORMAT(OPCODE_JNS,  off)]
-def JS  (off):         return [C_FORMAT(OPCODE_JS,   off)]
-def JNC (off):         return [C_FORMAT(OPCODE_JNC,  off)]
-def JC  (off):         return [C_FORMAT(OPCODE_JC,   off)]
-def JNO (off):         return [C_FORMAT(OPCODE_JNO,  off)]
-def JO  (off):         return [C_FORMAT(OPCODE_JO,   off)]
-
-def JNE (off):         return [C_FORMAT(OPCODE_JNE,  off)]
-def JE  (off):         return [C_FORMAT(OPCODE_JE,   off)]
-def JUGE(off):         return [C_FORMAT(OPCODE_JUGE, off)]
-def JULT(off):         return [C_FORMAT(OPCODE_JULT, off)]
-def JUGT(off):         return [C_FORMAT(OPCODE_JUGT, off)]
-def JULE(off):         return [C_FORMAT(OPCODE_JULE, off)]
-def JSGE(off):         return [C_FORMAT(OPCODE_JSGE, off)]
-def JSLT(off):         return [C_FORMAT(OPCODE_JSLT, off)]
-def JSGT(off):         return [C_FORMAT(OPCODE_JSGT, off)]
-def JSLE(off):         return [C_FORMAT(OPCODE_JSLE, off)]
-
-def ILL(insn):         return [insn & 0xffff]
-
-def MOVI(rd, imm16):
-    assert imm16 in range(65536)
-    if imm16 in range(256):
-        return MOVL(rd, imm16)
-    else:
-        return MOVH(rd, (imm16 >> 8) + ((imm16 >> 7) & 1)) + \
-               [I_FORMAT(OPCODE_ADDI, rd, imm16 & 0xff, u=True)]
+    @parse.with_pattern(r"R\d+")
+    @classmethod
+    def from_str(cls, input):
+        return super().from_str(input)
 
 
-def XCHG(rx, ry):
-    assert rx != ry
-    return [
-        *XOR(rx, rx, ry),
-        *XOR(ry, ry, rx),
-        *XOR(rx, rx, ry)
-    ]
+class Imm(mc.Operand):
+    format     = "{:#x}"
+    bits       = abc.abstractproperty()
+    values     = abc.abstractproperty()
 
-
-def L(label): return label
-
-def assemble(code):
-    flat_code = []
-    labels    = {}
-    for elem in code:
-        if isinstance(elem, str):
-            assert elem not in labels
-            labels[elem] = len(flat_code)
+    @classmethod
+    def prepare(cls, value):
+        if hasattr(value, "__int__"):
+            value = int(value)
         else:
-            flat_code += elem
-    for offset, elem in enumerate(flat_code):
-        if isinstance(elem, types.LambdaType):
-            flat_code[offset] = elem(lambda label: labels[label] - offset - 1)
-    return flat_code
+            try:
+                value = int(value, 0)
+            except ValueError:
+                return value # will be resolved later
+        if value not in cls.values:
+            raise ValueError(f"Immediate operand {value} must be in range "
+                             f"{cls.values.start}..{cls.values.stop}")
+        return value
+
+    @property
+    def is_legal(self):
+        return self.value & 0xffff not in range((+1 << self.bits - 1) & 0xffff,
+                                                (-1 << self.bits - 1) & 0xffff)
+
+    @classmethod
+    def from_str(cls, input):
+        return cls(input)
+
+    @classmethod
+    def value_from_bits(cls, input):
+        return _signed(input, cls.bits)
+
+    def __str__(self):
+        if hasattr(self.value, "__int__"):
+            return self.format.format(self.value)
+        return self.value
+
+    def __call__(self, resolver):
+        if hasattr(self.value, "__int__"):
+            return self # never relocate literals
+        rel_value = resolver(self.value)
+        if rel_value is None:
+            return self # undefined reference
+        return type(self)(rel_value)
+
+
+class ImmLUT(Imm):
+    lut_to_imm = abc.abstractproperty()
+    imm_to_lut = abc.abstractproperty()
+
+    @property
+    def is_legal(self):
+        return (self.value & 0xffff) in self.imm_to_lut
+
+    @classmethod
+    def value_from_bits(cls, input):
+        return cls.lut_to_imm[input]
+
+    @classmethod
+    def bits_from_value(cls, value):
+        # For legal immediates, look up the values in the inverse LUT.
+        # For illegal immediates, trim the value to the part that fits into the instruction;
+        # the rest of the immediate will be handled in `Instr.encode()`.
+        return cls.imm_to_lut.get(value & 0xffff, value & ((1 << cls.bits) - 1))
+
+
+class Imm3AL(ImmLUT):
+    bits       = 3
+    values     = range(-1 << 15, 1 << 16)
+    lut_to_imm = [0x0000, 0x0001, 0x8000, 0x1234, # FIXME
+                  0x00ff, 0xff00, 0x7fff, 0xffff]
+    imm_to_lut = {v: k for k, v in enumerate(lut_to_imm)}
+
+class Imm3SR(ImmLUT):
+    bits       = 3
+    values     = range(0, 16)
+    lut_to_imm = [8, 1, 2, 3, 4, 5, 6, 7]
+    imm_to_lut = {v: k for k, v in enumerate(lut_to_imm)}
+
+class Imm5 (Imm):
+    bits       = 5
+    values     = range(-1 << 15, 1 << 16)
+
+class Imm8 (Imm):
+    bits       = 8
+    values     = range(-1 << 15, 1 << 16)
+
+class Imm13(Imm):
+    bits       = 13
+    values     = range(-1 << bits, 1 << bits)
+
+    @property
+    def is_legal(self):
+        return self.value in range(-1 << self.bits - 1, 1 << self.bits)
+
+    @classmethod
+    def value_from_bits(cls, input):
+        return input
+
+
+class Instr(mc.Instr):
+    abbrevs    = {"D": "rsd", "A": "ra", "B": "rb", "i": "imm"}
+    formats    = {"R": Reg, "I3AL": Imm3AL, "I3SR": Imm3SR, "I5": Imm5, "I8": Imm8, "I13": Imm13}
+    mnemonics  = {}
+    decodings  = {}
+
+    _i3_mask   = 0b0000000000000_111
+    _i13_mask  = 0b000_1111111111111
+    _ext_code  = 0b110_0000000000000
+    _ext_mask  = 0b111_0000000000000
+
+    @property
+    def max_length(self):
+        if hasattr(self, "imm"):
+            return 2
+        return 1
+
+    @property
+    def length(self):
+        if hasattr(self, "imm") and not self.imm.is_legal:
+            return 2 # EXTI plus instruction
+        return 1 # just the instruction
+
+    def encode(self, stream, use_longest=False):
+        # Translate the instruction first, so that we don't append anything to the stream if there
+        # is an unresolved relocation.
+        instr_encoding = int(self)
+        length = 1
+        # If the instruction has an immediate...
+        if hasattr(self, "imm"):
+            # ... which is too large to fit into the field in the encoding, or might participate
+            # in relocation later...
+            if use_longest or not self.imm.is_legal:
+                # ... then encode an EXTI instruction first.
+                stream.append(self._ext_code | ((self.imm.value >> 3) & self._i13_mask))
+                length += 1
+        # And encode the instruction in either case.
+        stream.append(instr_encoding)
+        return length
+
+    @classmethod
+    def decode(cls, stream, index=0):
+        # If there's an EXTI prefix...
+        if stream[index] & cls._ext_mask == cls._ext_code:
+            ext_imm = stream[index] & cls._i13_mask
+            # ... followed by a non-EXTI instruction...
+            if len(stream) > index + 1 and (stream[index + 1] & cls._ext_mask != cls._ext_code):
+                # ... then decode two opcodes as one instruction with a large immediate...
+                instr   = cls.from_int(stream[index + 1])
+                ext_imm = _signed((ext_imm << 3) | (stream[index + 1] & cls._i3_mask), 16)
+                if hasattr(instr, "imm"):
+                    # ... if it takes an immediate at all.
+                    instr.imm.value = ext_imm
+                    return instr, 2
+        # Othrewise, decode one opcode as one instruction, as usual.
+        return cls.from_int(stream[index]), 1
