+++ examples/software/Makefile
