--- formal/fsm_formal.sby
@@ -1,15 +1,0 @@
-[options]
-mode bmc
-depth 16
-
-[engines]
-smtbmc
-
-[script]
-read_verilog fsm_core_fi.v
-read_verilog -formal formal.sv
-prep -top boneless_formal
-
-[files]
-fsm_core_fi.v
-formal.sv
